Updated Turkish UI and refactored token retrieval
Updated the UI in `EmployeeInfo.cshtml` to Turkish and refactored the method of retrieving the authentication token in `EmployeeInfo.cshtml` and `Index.cshtml` to use a new function `getCookie(name)`. The URL for fetching employee data has been updated in `EmployeeInfo.cshtml`. Added additional console logs in `Index.cshtml` for debugging. Created a new notifications page `Notifications.cshtml` that displays notifications using a model `List<ResultNotificationDto>`. Added a function `ConvertUnixTimestampToDateTime(long unixTimestamp)` to convert Unix timestamps to a readable date-time format. The layout of the notifications page includes a card for each notification, displaying the time and message of the notification.