`Update AddDepartment.cshtml and related changes`
` `

`This commit includes several updates to the AddDepartment.cshtml file and related changes. The authorization check for the "AdminPolicy" has been removed and the error message for users without access to the admin page has been deleted. The form now uses the "AddDepartment" action and the POST method, and the JavaScript code for submitting the form and making a POST request to the API has been removed. The text in the header and labels has been changed from Turkish to English. The "Index" action in the back button has been changed to "DepartmentList". Lastly, the "_ValidationScriptsPartial" partial view has been added to the Scripts section.`