Updated LoginAdminController and refactored login process
Updated the `LoginAdminController.cs` file with several significant changes. The `FluentValidation.Results` namespace was removed and replaced with the `Microsoft.AspNetCore.Authorization` namespace. Other namespaces were also rearranged, added, or removed. The `Login` method signature was updated to `LoginAdmin`. The URL in the `PostAsJsonAsync` method was updated to point to the Azure website instead of localhost. For users with the "Employee" role, the `AuthenticationToken` cookie is now deleted and the user is redirected to the "AccessDenied" action in the "ErrorPage" controller. Lastly, a large block of commented-out code handling the admin login process was removed.