Updated URL endpoint in LoginEmployeeController
Updated the URL endpoint for the `PostAsJsonAsync` method in the `LoginEmployee` function of the `LoginEmployeeController.cs` file. The POST request now points to the remote server "https://trackingprojectwebappservice20240505190044.azurewebsites.net/api/Auth/Login" instead of the local server. This change indicates a shift from a local development environment to a production environment hosted on Azure.