`Update routes and authorization in AnnouncementController`
Changes made to the `AnnouncementController` class include the removal of the `[Authorize(Policy = "AdminPolicy")]` and `[Route("/Announcement/{otherId}")]` attributes, meaning the controller is no longer restricted to admin users and the route to access the controller has been changed. The `[HttpGet("Add")]` and `[HttpPost("Add")]` attributes in the `AddAnnouncement()` methods have been changed to `[HttpGet]` and `[HttpPost]` respectively, altering the routes to access and post data to these methods. The class is using the `UI_Layer.Models` namespace and is located in the `UI_Layer.Controllers.Admin` namespace.